+++ AcademicProject/ApiAcademic/Controllers/SubscriberController.cs
@@ -25,25 +25,58 @@ namespace ApiAcademic.Controllers
         // GET api/subscriber/5
         public Subscriber Get(long id)
         {
+            if (id == 0)
+                throw new HttpResponseException(HttpStatusCode.BadRequest);
+            
             return _subscriberrepository.getSubscriberById(id);
         }
 
         // POST api/subscriber
         public long Post([FromBody]Subscriber subscriber)
         {
+            if(!ValidateSubscriber(subscriber))
+                throw new HttpResponseException(HttpStatusCode.BadRequest);
+
            return _subscriberrepository.SaveSubscriber(subscriber);
         }
 
         // PUT api/subscriber/5
         public long Put([FromBody]Subscriber subscriber)
         {
+            if (!ValidateSubscriber(subscriber))
+                throw new HttpResponseException(HttpStatusCode.BadRequest);
+
             return _subscriberrepository.SaveSubscriber(subscriber);
         }
 
         // DELETE api/subscriber/5
         public bool Delete(int id)
         {
+            if (id==0)
+                throw new HttpResponseException(HttpStatusCode.BadRequest);
+
             return _subscriberrepository.DeleteSubscriber(id);
+        }
+
+        public bool ValidateSubscriber(Subscriber subscriber)
+        {
+            bool result = true;
+
+            if (subscriber.id == 0) result = false;
+
+            if (subscriber.lastname == null || subscriber.lastname == string.Empty) result = false;
+
+            if (subscriber.firstname == null || subscriber.firstname == string.Empty) result = false;
+
+            if (subscriber.email == null || subscriber.email == string.Empty) result = false;
+
+            if (subscriber.password == null || subscriber.password == string.Empty) result = false;
+
+            if (subscriber.school == null || subscriber.school == string.Empty) result = false;
+
+            if (subscriber.username == null || subscriber.username == string.Empty) result = false;
+
+            return result;
         }
     }
 }
